Fixed installation issues for Vagrant system
Deprecated the use of robobsd.sh to revert back to the original
nanobsd.sh, as I really don't want to have to write my own. As
a result, I cleaned up some of the configuration files and added
support for some of the RoboBSD concepts in the .nano file. Also
added /etc/group and /etc/passwd files, along with packages and
sudo configuration files.